#16 Refactor project structure
-Separate CMake directories created
-Now .h files for each module are in include/opty for unification
-Linking static libraries for each module to your projects is possible
