+++ app/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_executable(app main.cpp)
+
+target_link_libraries(app PUBLIC Numerical GradOpt Newton UniOpt)
