+++ src/CMakeLists.txt
@@ -1,0 +1,4 @@
+add_subdirectory(UniOpt)
+add_subdirectory(Newton)
+add_subdirectory(Numerical)
+add_subdirectory(GradOpt)
