+++ src/Numerical/include/opty/linalg.h
