+++ src/Numerical/include/opty/vec.h
